Simplify updating of lower_sum value
Simplify the update of the lower_sum field in the timer stack by
shifting the contents by one struct: The value for the current timer
is now in a dedicated variable, the field in the struct holds the
value for the previous entry on the timer stack. This removes the need
to access the previous entry when a timer is stopped.