Attempt to estimate the timing overhead
Add an estimate for the timing overhead to the output file. It is
calculated by starting/stopping a timer 10 or 1000 times before and
after R itself runs. The average number of measurement units per timer
start are shown as "OverheadEstimate" in the output file. The values
are different because the 1000 iteration measurement profits a lot
more from the CPU cache than the 10 iteration measurement, so they are
shown as a value range.