Shuffle Eval-related timers around
Remove the TR_Eval, TR_bcEval and TR_dotBuiltin timers (for now) - the
expected overhead for these timers is rather large compared to the
runtime accumulated on them, resulting in a skewed result compared to
C-level profiling tools because the functions are called rather
often. In order to estimate the run time needed for evaluating
parameter lists of builtins, evalList has received its own timer instead.